requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.7
autopep8==2.3.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20241021
dill==0.3.9
Django==5.1.2
django-celery-beat==2.7.0
django-filter==24.3
django-import-export==4.3.3
django-simple-history==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
flake8==7.1.1
flower==2.0.1
humanize==4.11.0
inflection==0.5.1
isort==5.13.2
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.2
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.1
tablib==3.7.0
tomlkit==0.13.2
tornado==6.4.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
